Moved specific inspection files under SolderPasteInspection namespace
