+++ .gitignore
@@ -29,3 +29,4 @@
 # Uncomment if necessary however generally it will be regenerated when needed
 #!**/packages/repositories.config
 /CFXUnitTests/docs
+/CFX/ClassDiagram1.cd
